--- src/Core/Query/GetParameters/index.ts
@@ -1,1 +1,0 @@
-export * as GetParameters from "./GetParameters";
