+++ src/Core/Query/index.ts
@@ -1,4 +1,1 @@
-import { GetParameters } from "./GetParameters";
-
-export { GetParameters };
 export * as Query from "./Query";
