+++ src/Data/Managers/index.ts
@@ -37,6 +37,5 @@ export * from "./PromoteVersion";
 export * from "./GetVersionResources";
 export * from "./ControlAgent";
 export * from "./GetCompilerStatus";
-export * from "./GetParameters";
 export * from "./GetDesiredStateDiff";
 export * as GetVersionedResourceDetails from "./GetVersionedResourceDetails";
