+++ src/Slices/Parameters/Core/Route.ts
@@ -1,0 +1,11 @@
+import { Route } from "@/Core";
+
+export const path = "/parameters";
+
+export const route = (base: string): Route<"Parameters"> => ({
+  kind: "Parameters",
+  parent: "Home",
+  path: `${base}${path}`,
+  generateLabel: () => "Parameters",
+  environmentRole: "Required",
+});
