+++ src/Slices/Parameters/index.ts
@@ -1,0 +1,1 @@
+export * as Parameters from "./Core/Route";
