+++ src/Test/Data/index.ts
@@ -16,6 +16,5 @@ export * as EnvironmentSettings from "./EnvironmentSettings";
 export * as Environment from "./Environment";
 export * as ServiceInventory from "./ServiceInventory";
 export * as DesiredStateVersions from "./DesiredStateVersions";
-export * as Parameters from "./Parameters";
 export * as DesiredStateDiff from "./DesiredStateDiff";
 export * as VersionedResourceDetails from "./VersionedResourceDetails";
