+++ src/UI/Pages/PrimaryPageManager.tsx
@@ -9,6 +9,7 @@ import { DiagnosePage } from "@S/Diagnose/UI";
 import { EventsPage } from "@S/Events/UI";
 import { FactsPage } from "@S/Facts/UI";
 import { NotificationCenterPage } from "@S/Notification/UI/Center";
+import { ParametersPage } from "@S/Parameters/UI";
 import { ResourcesPage } from "@S/Resource/UI/ResourcesPage";
 import { ServiceInstanceHistoryPage } from "@S/ServiceInstanceHistory/UI";
 import { SettingsPage } from "@S/Settings/UI";
@@ -20,7 +21,6 @@ import { DesiredStateDetailsPage } from "./DesiredStateDetails";
 import { DesiredStateResourceDetailsPage } from "./DesiredStateResourceDetails";
 import { EditInstancePage } from "./EditInstance";
 import { HomePage } from "./Home";
-import { ParametersPage } from "./Parameters";
 import { ResourceDetailsPage } from "./ResourceDetails";
 import { ServiceCatalogPage } from "./ServiceCatalog";
 import { ServiceInventoryPage } from "./ServiceInventory";
